refactor: move hooks directory out of router
