+++ app/(site)/components/PageContent.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import useOnPlay from '~/app/hooks/useOnPlay';
 import SongItem from '~/components/SongItem';
+import useOnPlay from '~/hooks/useOnPlay';
 import { Song } from '~/types';
 
 interface PageContentProps {
@@ -15,9 +15,7 @@ const PageContent: React.FC<PageContentProps> = ({ songs }) => {
     return <div className="mt-4 text-neutral-400">No songs available.</div>;
   }
   return (
-    <div
-      className="grid grid-cols-2 gap-4 mt-4  sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-8"
-    >
+    <div className="grid grid-cols-2 gap-4 mt-4  sm:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-8">
       {songs.map((item) => (
         <SongItem
           key={item.id}
