+++ app/liked/components/LikedContent.tsx
@@ -3,10 +3,10 @@
 import { useRouter } from 'next/navigation';
 import { useEffect } from 'react';
 
-import useOnPlay from '~/app/hooks/useOnPlay';
-import { useUser } from '~/app/hooks/useUser';
 import LikeButton from '~/components/LikeButton';
 import MediaItem from '~/components/MediaItem';
+import useOnPlay from '~/hooks/useOnPlay';
+import { useUser } from '~/hooks/useUser';
 import { Song } from '~/types';
 
 interface LikedContentProps {
@@ -26,9 +26,7 @@ const LikedContent: React.FC<LikedContentProps> = ({ songs }) => {
 
   if (songs.length === 0) {
     return (
-      <div
-        className="flex flex-col w-full px-6  gap-y-2 text-neutral-400"
-      >
+      <div className="flex flex-col w-full px-6  gap-y-2 text-neutral-400">
         No liked songs.
       </div>
     );
