+++ components/Player.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import useGetSongById from '~/app/hooks/useGetSongById';
-import useLoadSongUrl from '~/app/hooks/useLoadSongUrl';
-import usePlayer from '~/app/hooks/usePlayer';
+import useGetSongById from '~/hooks/useGetSongById';
+import useLoadSongUrl from '~/hooks/useLoadSongUrl';
+import usePlayer from '~/hooks/usePlayer';
 import PlayerContent from './PlayerContent';
 
 const Player = () => {
