+++ components/SearchInput.tsx
@@ -4,7 +4,7 @@ import { useRouter } from 'next/navigation';
 import qs from 'query-string';
 import { useEffect, useState } from 'react';
 
-import useDebounce from '~/app/hooks/useDebounce';
+import useDebounce from '~/hooks/useDebounce';
 import Input from './Input';
 
 const SearchInput = () => {
