+++ components/SubscribeModal.tsx
@@ -3,8 +3,8 @@
 import { useState } from 'react';
 import toast from 'react-hot-toast';
 
-import useSubscribeModal from '~/app/hooks/useSubscribeModal';
-import { useUser } from '~/app/hooks/useUser';
+import useSubscribeModal from '~/hooks/useSubscribeModal';
+import { useUser } from '~/hooks/useUser';
 import { postData } from '~/libs/helpers';
 import { getStripe } from '~/libs/stripeClient';
 import { Price, ProductWithPrice } from '~/types';
